requirements.txt
fastapi
uvicorn[standard]
openai
anthropic
python-multipart
firebase-admin
aiosmtplib
python-dotenv
reportlab
